Move form builder to separate file
